Modification id de long à int + méthode findFilmById

+++ src/main/java/fr/eni/tp/filmotheque/bll/FilmService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface FilmService {
     List<Film> consulterFilms();
 
-    Film consulterFilmParId(long id);
+    Film consulterFilmParId(int id);
 
     //List<Genre> consulterGenres();
 
@@ -16,7 +16,7 @@ public interface FilmService {
 
     //Genre consulterGenreParId(long id);
 
-    Participant consulterParticipantParId(long id);
+    Participant consulterParticipantParId(int id);
 
     void creerFilm(Film film);
 }
