+++ src/main/java/fr/eni/tp/filmotheque/bo/Genre.java
@@ -9,7 +9,7 @@ public class Genre implements Serializable {
 	 */
 	private static final long serialVersionUID = 1L;
 	//Attributs
-	private long id;
+	private int id;
 	private String titre;
 
 	//Default Constructor
@@ -21,17 +21,17 @@ public class Genre implements Serializable {
 		this.titre = titre;
 	}
 
-	public Genre(long id, String titre) {
+	public Genre(int id, String titre) {
 		this.id = id;
 		this.titre = titre;
 	}
 	
 	//Getter + Setter
-	public long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
